Adds background image, changes text to italian restaurent
The image is nice but clearly italian, so I changed from a French
restaurent (I couldn't find a nice equivalent image for French cuisine in
the 10 mins I spent looking...)

Also added some containers so the spacing works, e.g. the image is 100
vh and the text is size restricted. And the colour is now white.

Added a link too for photo attribution - not yet at the bottom of the
page.